@@ -13,24 +13,24 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/animejs": "^3.1.12",
-		"@types/three": "^0.160.0",
-		"sass": "^1.69.7",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.12"
+		"@sveltejs/adapter-auto": "^6.0.0",
+		"@sveltejs/kit": "^2.20.8",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@types/animejs": "^3.1.13",
+		"@types/three": "^0.176.0",
+		"sass": "^1.87.0",
+		"svelte": "^5.28.2",
+		"svelte-check": "^4.1.6",
+		"tslib": "^2.8.1",
+		"typescript": "^5.8.3",
+		"vite": "^6.3.4"
 	},
 	"dependencies": {
-		"animejs": "^3.2.1",
+		"animejs": "^4.0.2",
 		"bezier-easing": "^2.1.0",
-		"detect-gpu": "^4.0.7",
+		"detect-gpu": "^5.0.70",
 		"slickscrolljs": "^2.0.2",
-		"three": "^0.136.0"
+		"three": "^0.176.0"
 	},
 	"type": "module"
 }
