@@ -10,23 +10,24 @@
 	import type { WorkData } from '$lib/types';
 
 	// JSON Work data fetched from the work-data.json file
-	let data: WorkData = [];
+	let data: WorkData = $state([]);
 	workItemsFetch.subscribe(val => {
 		if (val !== undefined) data = val;
 	});
 
 	// Is user holding click
-	let isMouseDown: boolean = false;
+	let isMouseDown: boolean = $state(false);
 	isWorkScroll.subscribe(val => isMouseDown = val);
 
 
-	let workContainer: HTMLElement;
-	let container: HTMLElement, listContainer: HTMLElement; // Containers for Three meshes
-	let images: HTMLImageElement[] = []; // Array of images to be passed to WebGL Shader
-	let workItems: HTMLElement[]  = []; // Array of workItems to be animated
+	let workContainer: HTMLElement;// = $state()!;
+	let container: HTMLElement;// = $state()!; 
+	let listContainer: HTMLElement;// = $state()!; // Containers for Three meshes
+	let images: HTMLImageElement[] = [];// = $state([]); // Array of images to be passed to WebGL Shader
+	let workItems: HTMLElement[] = [];//  = $state([]); // Array of workItems to be animated
 
-	let breakTitleWords: boolean = false;
-	let currentActive: number = -1; // The work item to expand
+	let breakTitleWords: boolean = $state(false);
+	let currentActive: number = $state(-1); // The work item to expand
 
 	let inViewResolve: (_: boolean) => void;
 	const inViewPromise: Promise<boolean> = new Promise((resolve) => {
@@ -46,6 +47,7 @@
 		lerpSpeed = 0.1;
 
 		onHold = (e: MouseEvent) => {
+			e.preventDefault();
 			if (currentActive >= 0 || isMouseDown || (e.target! as HTMLElement).classList.contains("button")) return;
 
 			this.initialMouseX = e.clientX;
@@ -65,6 +67,7 @@
 		}
 	
 		onMouseMove = (e: MouseEvent) => {
+			e.preventDefault();
 			if (!isMouseDown) return; 
 			this.currentMouseX = e.clientX;
 
@@ -141,10 +144,10 @@
 	<div class="content-wrapper" 
 		role="listbox"
 		tabindex="0"
-		on:mousedown|preventDefault={slider.onHold}
-		on:mouseup={slider.onRelease}
-		on:mouseleave={slider.onRelease}
-		on:mousemove|preventDefault={slider.onMouseMove}
+		onmousedown={slider.onHold}
+		onmouseup={slider.onRelease}
+		onmouseleave={slider.onRelease}
+		onmousemove={slider.onMouseMove}
 		bind:this={container}
 		class:disabled={currentActive >= 0}
 		use:workListIntro={{ promise: inViewPromise, onComplete: async () => {
@@ -168,7 +171,7 @@
 
 							<div class="img-wrapper">
 								{#await loadImage(`assets/imgs/work-back/${item.id}/cover.jpg`) then src}
-									<img bind:this={images[i]} src="{src}" on:dragstart|preventDefault draggable="false" alt="{item.title} Background">
+									<img bind:this={images[i]} src="{src}" ondragstart={(e) => {e.preventDefault()}} draggable="false" alt="{item.title} Background">
 								{/await}
 							</div>
 							{#await inViewPromise then _}
@@ -198,7 +201,7 @@
 
 									<button 
 										class="button item-link interactive"
-										on:click={() => toggleActiveItem(i)}
+										onclick={() => toggleActiveItem(i)}
 										in:maskSlideIn={{
 											duration: 900,
 											delay: (i*30)+450,
@@ -241,12 +244,12 @@
 							use:titleSlide
 							out:letterSlideOut
 							class:breakTitleWords
-							on:introend={() => setTimeout(() => breakTitleWords = true, 100)}
-							on:outrostart={() => setTimeout(() => breakTitleWords = false, 100)}>
+							onintroend={() => setTimeout(() => breakTitleWords = true, 100)}
+							onoutrostart={() => setTimeout(() => breakTitleWords = false, 100)}>
 
 							{data[currentActive].title}
 						</h1>
-						<button class="close-button-wrapper interactive" on:click={() => toggleActiveItem(currentActive)}>
+						<button class="close-button-wrapper interactive" onclick={() => toggleActiveItem(currentActive)}>
 							<div 
 								class ="close-button"
 								in:maskSlideIn={{ reverse: true }} 
@@ -294,8 +297,8 @@
 
 <style lang="sass">
 
-@import "../consts"
-@include textStyles()
+@use "../consts" as consts
+@include consts.textStyles()
 
 #content-container.work-click-area
 	margin-top: 30vh
@@ -354,13 +357,13 @@
 
 					h6.caption
 						position: relative
-						font-family: $font
+						font-family: consts.$font
 						text-transform: uppercase
 						font-weight: normal
 						font-size: 1.9vh
 
 					.index
-						font-family: $font
+						font-family: consts.$font
 						position: relative
 						font-size: 2.1vh
 
@@ -378,7 +381,7 @@
 
 				h1.title
 					position: relative
-					font-family: $titleFont
+					font-family: consts.$titleFont
 					font-size: 7vw
 					text-transform: lowercase
 					font-weight: normal
@@ -435,7 +438,7 @@
 					.descriptor
 						line-height: 270%
 						letter-spacing: 0.5vh
-						font-family: $font
+						font-family: consts.$font
 						text-transform: uppercase
 						font-weight: normal
 						font-size: 1.4vh
@@ -447,7 +450,7 @@
 						flex-direction: column
 
 						li
-							font-family: $font
+							font-family: consts.$font
 							text-transform: uppercase
 							font-weight: normal
 							font-size: 1.7vh
@@ -583,7 +586,7 @@
 					font-size: 1vw
 					letter-spacing: 0.1vw
 					text-transform: uppercase
-					font-family: $font
+					font-family: consts.$font
 
 			.text-wrapper
 				box-sizing: border-box
@@ -603,7 +606,7 @@
 					text-transform: uppercase
 
 				.item-title
-					font-family: $font
+					font-family: consts.$font
 					font-weight: normal
 					font-size: 2.5vw
 					z-index: 0
